+++ docs/info/change_log.rst
@@ -7,7 +7,7 @@ Known issues
 __________________
 
 
-Version 0.45
+Version 0.4.1.000 2023-12-23
 _________________________
 
 Faster and simplified internal structure
@@ -15,7 +15,9 @@ Faster and simplified internal structure
 New features
 --------------------
 
-#. Most expensive step in 3D finding vertical cell is not 5 times faster, so net 2-3 times faster, as by default regrids  3D data to uniform S grid, to speed vertical cell search by factor of 5. Original native vertical grid short vertical walk search still available.
+#. Most expensive step in 3D finding vertical cell is not 5 times faster, so net 2-3 times faster, as by default regrids  3D data to uniform Sigma grid, to speed vertical cell search by factor of 5. Original native vertical grid short vertical walk search still available.
+#. By default uses diffusivity vertical profile for A_Z in random walk if variable mapped to A_Z_profile is found in the hydro file
+#. Uses bottom stress to calculate friction velocity field if variable mapped to bottom_stress is found in the hydro file
 #. Floating particle trajectory  modifier, forces particles to follow free surface
 #. Now exits if any the release group start time is outside time range of hindcast